Fix some stuff
Fixed the gravity, now the bodies are attracted towards the center of the sphere.
Added rotation to align the Y axis of the character with the center of the sphere.
Added camera rotation and movement.
TODO: Fix some parameters, finish uo the camera, it's not working perfectly, but it does the job more or less